server/requirements.txt
amqp==5.1.1
argcomplete==1.10.3
asgiref==3.5.2
async-timeout==4.0.2
billiard==3.6.4.0
celery==5.2.7
certifi==2022.9.24
cffi==1.15.1
chardet==3.0.4
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
compressed-rtf==1.0.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.3
defusedxml==0.7.1
Django==4.1.3
gunicorn==20.1.0
django-allauth==0.51.0
django-cors-headers==3.13.0
django-debug-toolbar==3.8.1
django-environ==0.9.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-yasg==1.21.4
ebcdic==1.1.1
extract-msg==0.28.7
idna==3.4
IMAPClient==2.1.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
lxml==4.9.2
MarkupSafe==2.1.1
oauthlib==3.2.2
olefile==0.46
packaging==23.0
pdfminer.six==20191110
Pillow==9.3.0
prompt-toolkit==3.0.36
psycopg2-binary==2.9.5
pycparser==2.21
pycryptodome==3.17
PyJWT==2.6.0
python3-openid==3.2.0
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
redis==4.4.2
requests==2.28.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
six==1.12.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sortedcontainers==2.4.0
soupsieve==2.3.2.post1
SpeechRecognition==3.8.1
sqlparse==0.4.3
tzdata==2022.7
tzlocal==4.2
uritemplate==4.1.1
urllib3==1.26.13
vine==5.0.0
wcwidth==0.2.5
whitenoise==6.2.0
XlsxWriter==3.0.7